continue

+++ migrations/20220413192435_create_user_table.sql
@@ -2,17 +2,17 @@ create table "user"
 (
   id                 text      not null primary key,
   name               text      not null default 'unknown',
-  removed_playlists  int       not null default 0,
-  removed_collection int       not null default 0,
+  removed_playlists  bigint    not null default 0,
+  removed_collection bigint    not null default 0,
   status             text      not null default 'active',
   created_at         timestamp not null default current_timestamp,
   updated_at         timestamp not null default current_timestamp,
   playing_track      text      not null default '',
 
-  lyrics_checked     int not null default 0,
-  lyrics_genius      int not null default 0,
-  lyrics_musixmatch  int not null default 0,
-  lyrics_profane     int not null default 0
+  lyrics_checked     bigint not null default 0,
+  lyrics_genius      bigint not null default 0,
+  lyrics_musixmatch  bigint not null default 0,
+  lyrics_profane     bigint not null default 0
 );
 
 create unique index user_id_uindex on "user" (id);
